+++ application/src/main/java/pt/amane/application/video/update/UpdateVideoCommand.java
@@ -1,5 +1,6 @@
 package pt.amane.application.video.update;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 import java.util.Set;
 import pt.amane.domain.video.Resource;
@@ -9,7 +10,7 @@ public record UpdateVideoCommand(
     String title,
     String description,
     Integer launchedAt,
-    Double duration,
+    BigDecimal duration,
     Boolean opened,
     Boolean published,
     String rating,
@@ -29,7 +30,7 @@ public record UpdateVideoCommand(
       final String title,
       final String description,
       final Integer launchedAt,
-      final Double duration,
+      final BigDecimal duration,
       final Boolean opened,
       final Boolean published,
       final String rating,
@@ -62,7 +63,7 @@ public record UpdateVideoCommand(
       final String title,
       final String description,
       final Integer launchedAt,
-      final Double duration,
+      final BigDecimal duration,
       final Boolean opened,
       final Boolean published,
       final String rating,
